Extract functions to the separate file
